requirements.txt
tinydb==4.8.0
pytest==7.4.3
python-dotenv==1.0.0
pydantic==1.10.13
pytest-mock==3.12.0